Extract decorator types to shared file
